# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/squad/import": {
      "post": {
        "description": "Fetches the manager's last-locked picks from the public API and persists them. No credential is involved and none is accepted (D-013): the manager id is an input, not an identity. The pre-deadline unsaved squad is not public and is not returned. This is the one endpoint that calls upstream while you wait, so it carries a short timeout; a second call for the same manager and gameweek is served from GET /api/squad/{managerId}.",
        "operationId": "SquadController_import",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSquadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The imported squad, now persisted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SquadDto"
                        },
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "errorCode": {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No such manager id.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        },
                        "success": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "errorCode": {
                          "type": "string",
                          "enum": [
                            "MANAGER_NOT_FOUND"
                          ]
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "The squad contains a player this app has not synced.\n\nThe manager exists but has no readable picks yet — picks become public after a deadline.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        },
                        "success": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "errorCode": {
                          "type": "string",
                          "enum": [
                            "SQUAD_NOT_AVAILABLE_YET"
                          ]
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "The official API timed out or failed. Not the manager id's fault.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        },
                        "success": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "errorCode": {
                          "type": "string",
                          "enum": [
                            "FPL_UPSTREAM_UNAVAILABLE"
                          ]
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Import a squad from a public FPL manager id.",
        "tags": [
          "squad"
        ]
      }
    },
...
    "/api/squad/{managerId}": {
      "get": {
        "description": "Reads Postgres. Makes no upstream call.",
        "operationId": "SquadController_get",
        "parameters": [
          {
            "name": "managerId",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stored squad.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SquadDto"
                        },
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "errorCode": {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "This manager has never been imported.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        },
                        "success": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "errorCode": {
                          "type": "string",
                          "enum": [
                            "SQUAD_NOT_IMPORTED"
                          ]
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "A previously imported squad, from our own store.",
        "tags": [
          "squad"
        ]
      }
    },
...
  },
  "info": {
    "title": "fpl-backend",
    "description": "The only HTTP surface of the fantasy-premier-league stack. Every response is wrapped in the ApiResponse envelope by a global interceptor; `data` holds the payload and `errorCode` is the stable key to switch on. No authentication anywhere — the product reads the public FPL API only (decision D-013).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "local"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponseMetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "description": "Echoed from x-request-id, or generated per request."
          },
          "durationMs": {
            "type": "number",
            "description": "Server-side handling time in milliseconds."
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "dataAsOfGw": {
            "type": "number",
            "description": "Which gameweek's data produced this. Present on every response carrying model output."
          }
        },
        "required": [
          "requestId",
          "durationMs",
          "generatedAt"
        ]
      },
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string",
            "nullable": true,
            "description": "Stable machine-readable key on failure, e.g. MANAGER_NOT_FOUND. Null when success is true."
          },
          "meta": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiResponseMetaDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "errorCode",
          "meta"
        ]
      },
...
      "SquadPickDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "description": "Our internal player id (cuid), not the FPL element id."
          },
          "fplId": {
            "type": "number",
            "description": "The player's FPL element id."
          },
          "webName": {
            "type": "string",
            "example": "Haaland"
          },
          "position": {
            "type": "string",
            "enum": [
              "GKP",
              "DEF",
              "MID",
              "FWD"
            ]
          },
          "teamShortName": {
            "type": "string",
            "example": "MCI"
          },
          "nowCost": {
            "type": "number",
            "description": "Market price in tenths of a million. 55 is £5.5m."
          },
          "sellValue": {
            "type": "number",
            "nullable": true,
            "description": "What a sale would return, in tenths. NULL for an imported squad and that is correct: no public endpoint carries a purchase or selling price. Filled by the transfer planner (B-008)."
          },
          "slot": {
            "type": "number",
            "description": "1-11 starting XI, 12-15 bench in substitution order. Called `position` upstream, renamed here because a pick already has a position in the GKP/DEF/MID/FWD sense and one field cannot be both."
          },
          "multiplier": {
            "type": "number",
            "description": "1 normally, 2 for the captain, 3 under the triple-captain chip."
          },
          "isCaptain": {
            "type": "boolean"
          },
          "isViceCaptain": {
            "type": "boolean"
          }
        },
        "required": [
          "playerId",
          "fplId",
          "webName",
          "position",
          "teamShortName",
          "nowCost",
          "sellValue",
          "slot",
          "multiplier",
          "isCaptain",
          "isViceCaptain"
        ]
      },
      "SquadDto": {
        "type": "object",
        "properties": {
          "managerId": {
            "type": "number",
            "nullable": true,
            "description": "The FPL manager id this squad was imported from. NULL for the recommended squad, which belongs to nobody."
          },
          "managerName": {
            "type": "string",
            "nullable": true,
            "description": "The manager's display name, when the squad came from an import."
          },
          "gameweekId": {
            "type": "number",
            "description": "The gameweek this squad is the picks for."
          },
          "bank": {
            "type": "number",
            "description": "Money not spent, in tenths of a million."
          },
          "teamValue": {
            "type": "number",
            "description": "Squad value excluding the bank, in tenths of a million."
          },
          "activeChip": {
            "type": "string",
            "nullable": true,
            "description": "The chip active in this gameweek, e.g. bboost or 3xc. Null when none."
          },
          "source": {
            "type": "string",
            "enum": [
              "import",
              "recommended"
            ],
            "description": "Where this squad came from. The advice is identical either way."
          },
          "picks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SquadPickDto"
            }
          }
        },
        "required": [
          "managerId",
          "managerName",
          "gameweekId",
          "bank",
          "teamValue",
          "activeChip",
          "source",
          "picks"
        ]
      },
      "ImportSquadDto": {
        "type": "object",
        "properties": {
          "managerId": {
            "type": "number",
            "description": "A public FPL manager id — the number in a manager's /entry/<id>/ URL on the official site.",
            "example": 1,
            "minimum": 1
          }
        },
        "required": [
          "managerId"
        ]
      },
...
    }
  }
}